LPS-54861 Move to invoker package
