+++ shared/alloy-mvc-shared/src/com/liferay/alloy/mvc/jsonwebservice/AlloyControllerInvoker.java
@@ -12,30 +12,25 @@
  * details.
  */
 
-package com.liferay.alloy.mvc;
+package com.liferay.alloy.mvc.jsonwebservice;
 
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Inherited;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import javax.portlet.PortletRequest;
+import com.liferay.alloy.mvc.AlloyController;
+import com.liferay.alloy.mvc.AlloyPortlet;
+import com.liferay.portal.kernel.json.JSONSerializable;
+import com.liferay.portal.model.Portlet;
+import com.liferay.portal.theme.ThemeDisplay;
 
 /**
  * @author Ethan Bustad
  */
-@Documented
-@Inherited
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.METHOD})
-public @interface JSONWebServiceMethod {
+public interface AlloyControllerInvoker {
 
-	public String lifecycle() default PortletRequest.RENDER_PHASE;
+	public JSONSerializable invokeAlloyController(
+			String lifecycle, Object... parameters)
+		throws Exception;
 
-	public String[] parameterNames() default {};
-
-	public Class<?>[] parameterTypes() default {};
+	public void setProperties(
+		ThemeDisplay themeDisplay, AlloyPortlet alloyPortlet, Portlet portlet,
+		Class<? extends AlloyController> controllerClass, String controller);
 
 }
